Updated slurm training script for efficient local disk usage, checkpointing update

+++ gpu_run.sh
@@ -5,14 +5,15 @@
 
 # Set up sbatch arguments
 
-#SBATCH --job-name=PFNs1                         ## Name of the job.
+#SBATCH --job-name=resnet_ln                     ## Name of the job.
 #SBATCH -A kgreif                                 ## account to charge 
 #SBATCH -p free-gpu                               ## partition/queue name
 #SBATCH --gres=gpu:V100:1                         ## Use only 1 GPU
 #SBATCH --nodes=1                                 ## (-N) number of nodes to use
-#SBATCH --tmp=50G                                 ## Request 80GB of scratch to hold data
+#SBATCH --tmp=150G                                 ## Request 80GB of scratch to hold data
+#SBATCH --mem=10G
 
-#SBATCH --array=1-10
+#SBATCH --array=1-5
 
 #SBATCH --error=./outfiles/%x_%a.err       ## error log file
 #SBATCH --output=./outfiles/%x_%a.out      ## output file
@@ -21,43 +22,53 @@
 #SBATCH --mail-user=kgreif@uci.edu                ## to this address
 
 # For puposes of file copying, we want jobs to start running in 5 second increments
-waitcount=0
-while [ $waitcount -le ${SLURM_ARRAY_TASK_ID} ]
+waitcount=1
+while [ $waitcount -lt ${SLURM_ARRAY_TASK_ID} ]
 do
-	sleep 5
+	echo "Sleeping for 10 seconds"
+	sleep 10
 	waitcount=$(($waitcount + 1))	
 done
 
 # Setup environment
 echo "Building software environment"
-module load cuda/11.7.1
+module load cuda/11.4.0
 export XLA_FLAGS=--xla_gpu_cuda_data_dir=$CUDA_PATH
 
 # Print out hostname and date of job
 echo "Found a node, here's some info: "
 hostname; date
 echo $SLURM_JOB_NAME
-echo $SLURM_JOB_ID
+echo $SLURM_ARRAY_JOB_ID
 echo $SLURM_ARRAY_TASK_ID
 echo "================================"
 
 # Set tmpdir to correct location
-export TMPDIR=/tmp/tt_data
+export TMPDIR=/tmp/kgreif/${SLURM_ARRAY_JOB_ID}
 tempname="$TMPDIR/temp.h5"
 trainname="$TMPDIR/train.h5"
 
-# If temp data file does not exist, transfer file from disk to scratch
-if [ ! -f $tempname ]
+# If data file does not exist, transfer file from disk to scratch
+echo "Dealing with data transfer"
+
+if [ ! -f $tempname ] && [ ! -f $trainname ]
 then
 	echo "Now transferring file from /pub to /tmp"
 	mkdir -p $TMPDIR
-	cp /pub/kgreif/samples/h5dat/train_s2_ln_small.h5 $tempname
+	cp /pub/kgreif/samples/h5dat/train_s2.h5 $tempname
 	mv $TMPDIR/temp.h5 $trainname
 	ls $TMPDIR
-# If temp data does exist, wait until transfer finishes and then continue
+# If data does exist, wait until transfer finishes and then continue
 else
-	while [ ! -f $trainname ]; do sleep 30; done
+	echo "Found temp or train file from other jobs"
+	while [ ! -f $trainname ]
+       	do 
+		echo "Sleep for 30 seconds"
+		sleep 30
+	done
 fi
+echo "Transfer finished!"
+ls $TMPDIR
 
 # Set up directory tree
 homedir="/data/homezvol0/kgreif/toptag/ConstituentTopTagging"
@@ -70,23 +81,34 @@ mkdir -p ${logdir}
 
 # We want output files to sit in trdir, so make that working directory
 cd $trdir
-echo "In directory ${trdir}"
+echo "Now run training in directory ${trdir}"
 ls -lrth
 
 # Next build command to run python training script
-command="python ${homedir}/kf_train.py --numFolds 10 --fold ${SLURM_ARRAY_TASK_ID} --type pfn --phisizes 250 250 250 --fsizes 500 500 500 --latent_dropout 0.084 --dropout 0.036 -lr 7.9e-5 -b 256 --numEpochs 75 --maxConstits 80 --file ${trainname} --dir ${trdir} --logdir ${logdir}"
+command="python ${homedir}/kf_train.py --numFolds 10 --fold ${SLURM_ARRAY_TASK_ID} --type resnet --nodes 3 4 6 3 --dropout 0.5 --bnMom 0.1 --schedule 1 -lr 1e-2 -b 256 --numEpochs 20 --file ${trainname} --dir ${trdir} --logdir ${logdir}"
 
 # Run command
 echo "================================"
 echo "Will run command ${command}"
 $command
-echo -e "\nDone!"
+echo -e "\nDone with training!"
+
+# Delete tmp files if this is the last job on the node in this job array
+echo "Showing current status:"
+host=$(hostname)
+squeue -u kgreif -w $host -j ${SLURM_ARRAY_JOB_ID}
+running=$(squeue -u kgreif -w $host -j ${SLURM_ARRAY_JOB_ID} | wc -l)
+echo "Jobs running: ${running}"
+if [ $running -le 2 ]
+then
+	echo "Found this is the last running job for this array on this node"
+	rm -rf $TMPDIR
+else
+	echo "Other jobs in this array are still running on this node. Exiting"
+fi
 
 # Finally move output files from outfiles to trdir
 echo "================================"
 echo "Transferring output files..."
 mv ${homedir}/outfiles/${SLURM_JOB_NAME}_${SLURM_ARRAY_TASK_ID}.out ${trdir}
 mv ${homedir}/outfiles/${SLURM_JOB_NAME}_${SLURM_ARRAY_TASK_ID}.err ${trdir}
-
-# Delete mp files
-rm -rf $TMPDIR
